Clean up migration files and fix numbering
- Remove duplicate migrations 0026 and 0027
- Rename 0028 to 0026 to maintain sequential order
- Update journal to reflect cleaned up migrations

+++ drizzle/0026_volatile_goblin_queen.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "product_images" (
-	"id" bigserial PRIMARY KEY NOT NULL,
-	"product_id" bigint NOT NULL,
+	"id" serial PRIMARY KEY NOT NULL,
+	"product_id" integer NOT NULL,
 	"image_url" text NOT NULL,
 	"display_order" integer DEFAULT 0,
 	"is_primary" boolean DEFAULT false,
